feat: Move TOKEN_RISK to component file
