+++ apps/web/src/views/Swap/hooks/fetchTokenRisk.ts
@@ -2,16 +2,6 @@
 import { ChainId } from '@pancakeswap/sdk'
 import { ACCESS_RISK_API } from 'config/constants/endpoints'
 
-export const TOKEN_RISK = {
-  UNKNOWN: -1,
-  SOME_RISK: 0,
-  VERY_LOW: 1,
-  LOW: 2,
-  MEDIUM: 3,
-  HIGH: 4,
-  SIGNIFICANT: 5,
-} as const
-
 export interface RiskTokenInfo {
   address: string
   chainId: ChainId
